Fixed a profile bug
