+++ src/main/java/Main.java
@@ -8,6 +8,8 @@ import general_group.screens.GeneralGroupCreationScreen;
 import general_group.use_case.GeneralGroupCreateInteractor;
 import general_group.use_case.GeneralGroupCreateOutputBoundary;
 import general_group.use_case.GroupRepoInt;
+import get_friends.entities.Profile;
+import get_friends.entities.ProfileFactory;
 import get_friends.interface_adapters.GetFriendsController;
 import get_friends.interface_adapters.GetFriendsPresenter;
 import get_friends.use_case.GetFriendsInteractor;
@@ -16,6 +18,9 @@ import get_friends.use_case.GetFriendsOutputBoundary;
 import javax.swing.*;
 import java.awt.*;
 import java.io.IOException;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Main {
     public static void main(String[] args) {
@@ -49,12 +54,21 @@ public class Main {
                 genGroupOutputBoundary, groupFactory, profileData);
         GeneralGroupCreateController genGroupController = new GeneralGroupCreateController(genGroupInteractor);
 
+        //Testing with a random profile, must change later on
+        java.util.List<String> friendList = new ArrayList<>();
+        friendList.add("a");
+        friendList.add("b");
+        friendList.add("C");
+        List<String> emptyList = new ArrayList<>();
+        Profile testProfile = ProfileFactory.create("Danielle", "Dani", LocalDate.now(),
+                "", emptyList, emptyList, emptyList, emptyList, friendList, emptyList);
+
         GetFriendsOutputBoundary friendsOutputBoundary = new GetFriendsPresenter();
         GetFriendsInteractor friendsInteractor = new GetFriendsInteractor(friendsOutputBoundary);
         GetFriendsController friendsController = new GetFriendsController(friendsInteractor);
 
         GeneralGroupCreationScreen generalGroupCreationScreen = new GeneralGroupCreationScreen(genGroupController,
-                friendsController);
+                friendsController, testProfile);
         screens.add(generalGroupCreationScreen, "Welcome!");
         cardLayout.show(screens, "create");
         application.pack();
