+++ src/main/java/get_friends/entities/Profile.java
@@ -38,14 +38,6 @@ public class Profile{
         this.blockedUsers = blockedUsers;
     }
 
-    /**
-     * Testing purpose
-     */
-    public Profile(String userName, List<String> friends) {
-        this.userName = userName;
-        this.friends = friends;
-    }
-
     public String getUserName() {
         return this.userName;
     }
