+++ src/main/java/get_friends/entities/ProfileFactory.java
@@ -1,0 +1,22 @@
+package get_friends.entities;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public class ProfileFactory{
+    public static Profile create(String userName,
+                                                        String profileName,
+                                                        LocalDate dob,
+                                                        String description,
+                                                        List<String> socialLinks,
+                                                        List<String> sensitiveWords,
+                                                        List<String> interests,
+                                                        List<String> groups,
+                                                        List<String> friends,
+                                                        List<String> blockedUsers) {
+
+        return new Profile(
+                userName, profileName, dob, description, socialLinks, sensitiveWords, interests, groups, friends,
+                blockedUsers);
+    }
+}
